+++ applications/app-bemyplan/src/main/java/com/deploy/bemyplan/auth/controller/AuthController.java
@@ -11,9 +11,9 @@ import com.deploy.bemyplan.config.resolver.UserId;
 import com.deploy.bemyplan.controller.common.response.ResponseDTO;
 import com.deploy.bemyplan.domain.user.User;
 import com.deploy.bemyplan.domain.user.UserRepository;
-import com.deploy.bemyplan.service.user.UserService;
-import com.deploy.bemyplan.service.user.UserServiceUtils;
-import com.deploy.bemyplan.service.user.dto.request.CheckAvailableNameRequestDto;
+import com.deploy.bemyplan.user.service.UserService;
+import com.deploy.bemyplan.user.service.UserServiceUtils;
+import com.deploy.bemyplan.user.service.dto.request.CheckAvailableNameRequestDto;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
