+++ applications/app-bemyplan/src/main/java/com/deploy/bemyplan/auth/service/impl/GoogleAuthService.java
@@ -8,8 +8,8 @@ import com.deploy.bemyplan.domain.user.UserRepository;
 import com.deploy.bemyplan.domain.user.UserSocialType;
 import com.deploy.bemyplan.external.client.google.GoogleAuthApiClient;
 import com.deploy.bemyplan.external.client.google.dto.response.GoogleProfileInfoResponse;
-import com.deploy.bemyplan.service.user.UserService;
-import com.deploy.bemyplan.service.user.UserServiceUtils;
+import com.deploy.bemyplan.user.service.UserService;
+import com.deploy.bemyplan.user.service.UserServiceUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
