+++ applications/app-bemyplan/src/main/java/com/deploy/bemyplan/user/service/UserService.java
@@ -1,23 +1,15 @@
-package com.deploy.bemyplan.service.user;
+package com.deploy.bemyplan.user.service;
 
-import com.deploy.bemyplan.common.exception.model.NotFoundException;
-import com.deploy.bemyplan.domain.plan.Plan;
-import com.deploy.bemyplan.domain.plan.PlanRepository;
 import com.deploy.bemyplan.domain.user.User;
 import com.deploy.bemyplan.domain.user.UserRepository;
 import com.deploy.bemyplan.domain.user.WithdrawalUser;
 import com.deploy.bemyplan.domain.user.WithdrawalUserRepository;
-import com.deploy.bemyplan.service.user.dto.request.CheckAvailableNameRequestDto;
-import com.deploy.bemyplan.service.user.dto.request.CreateUserDto;
-import com.deploy.bemyplan.service.user.dto.response.CreatorInfoResponse;
+import com.deploy.bemyplan.user.service.dto.request.CheckAvailableNameRequestDto;
+import com.deploy.bemyplan.user.service.dto.request.CreateUserDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Objects;
-
-import static com.deploy.bemyplan.common.exception.ErrorCode.NOT_FOUND_PLAN_EXCEPTION;
 
 @RequiredArgsConstructor
 @Service
