+++ applications/app-bemyplan/src/main/java/com/deploy/bemyplan/user/service/dto/request/CheckAvailableNameRequestDto.java
@@ -1,0 +1,22 @@
+package com.deploy.bemyplan.user.service.dto.request;
+
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Pattern;
+
+@ToString
+@Getter
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class CheckAvailableNameRequestDto {
+
+    @NotBlank(message = "{user.nickname.notBlank}")
+    @Pattern(regexp = "^[ㄱ-ㅎ가-힣a-zA-Z0-9-_]{2,15}$", message = "{user.nickname.format}")
+    private String nickname;
+}
